+++ src/main/java/com/fzl/pojo/TClientLogExample.java
@@ -115,7 +115,7 @@ public class TClientLogExample {
             return (Criteria) this;
         }
 
-        public Criteria andClientIdEqualTo(Integer value) {
+        public Criteria andClientIdEqualTo(Long value) {
             addCriterion("client_id =", value, "clientId");
             return (Criteria) this;
         }
